Move work into NodePanel
